requirements.txt
flake8==3.7.8
flake8-docstrings==1.5.0
pytest==5.3.2
python-dateutil==2.8.0
requests==2.22.0
jsonschema==3.2.0
aiohttp==3.6.2
azure-functions>=1.2.0
nox==2019.11.9
furl==2.1.0
pytest-asyncio==0.10.0
